update: publicacion ahora con id UUID

+++ src/modules/publicacion/publicacion.service.ts
@@ -1,11 +1,14 @@
 import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { CreatePublicacionDto, UpdatePublicacionDto } from './dto/publicacion.dto';
-import { EventoService } from '../actividad/servicesAux/evento.service';
 import { InjectRepository } from '@nestjs/typeorm';
 import { PublicacionEntity } from './entities/publicacion.entity';
-import { EventoEntity } from 'src/evento/entities/evento.entity';
+import { EventoEntity } from '../../evento/entities/evento.entity'
 import { Repository } from 'typeorm';
 import { CategoriaService } from '../categoria/categoria.service';
+import { EventoService } from 'src/evento/evento.service';
+import { validate as validateUUID } from 'uuid';
+import {v4 as uuidv4} from 'uuid';
+
 
 @Injectable()
 export class PublicacionService {
@@ -34,23 +37,29 @@ export class PublicacionService {
     if (existePublicacion) {
       throw new HttpException({ message: 'La publicacion ya existe' }, HttpStatus.CONFLICT);
     }
+    const existeEvento = await this.eventoService.findOneById(createPublicacionDto.eventoId );
 
+    const idUUID = uuidv4();
+    const publicacionConUUID = Object.assign(createPublicacionDto, { id: idUUID } )
 
-    const existeEvento = await this.eventoService.findOneByName(createPublicacionDto.eventoNombre);
-    if (!existeEvento) {
-      throw new HttpException('Evento no encontrado', HttpStatus.NOT_FOUND);
-    }
+    // console.log('publicacionConUUID', publicacionConUUID);
 
-    const publicacionCompleta = Object.assign(createPublicacionDto, { eventoId : existeEvento._id });
-
-    console.log('publicacionCompleta', publicacionCompleta);
-
-    const datainto = this.publicacionRepository.create(publicacionCompleta);
+    const datainto = this.publicacionRepository.create(publicacionConUUID);
     return await this.publicacionRepository.save(datainto);
     
   }
 
   async findAll() {
+    const publicaciones = await this.publicacionRepository.find();
+
+    // const publicacionDetalles = await Promise.all(publicaciones.map(async (publi) => {
+    //   return await this.obtenerPublicacionesByIdEvento(publi);
+    // }));
+
+    return publicaciones;
+  }
+
+  async findAllDetailed() {
     const publicaciones = await this.publicacionRepository.find();
 
     const publicacionDetalles = await Promise.all(publicaciones.map(async (publi) => {
@@ -60,8 +69,17 @@ export class PublicacionService {
     return publicacionDetalles;
   }
 
-  async findOne(nombre: string) {
+  async findByName(nombre: string) {
+
     const publicacion = await this.publicacionRepository.findOne({ where: { tituloPublicacion: nombre } });
+    if (!publicacion) {
+      throw new HttpException({ message: 'Publicacion no encontrada' }, HttpStatus.NOT_FOUND);
+    }
+    return await this.obtenerPublicacionesByIdEvento(publicacion);
+  }
+
+  async findById(id: string) {
+    const publicacion = await this.publicacionRepository.findOne({ where: { id: id } });
 
     if (!publicacion) {
       throw new HttpException({ message: 'Publicacion no encontrada' }, HttpStatus.NOT_FOUND);
@@ -69,70 +87,79 @@ export class PublicacionService {
     return await this.obtenerPublicacionesByIdEvento(publicacion);
   }
 
-  async update(nombre: string, updPubliDto: UpdatePublicacionDto) {
+  async updateByName(nombre: string, updPubliDto: UpdatePublicacionDto) {
     const publicacion = await this.publicacionRepository.findOne({ where: { tituloPublicacion: nombre } });
 
     if (!publicacion) {
       throw new HttpException({ message: 'Publicacion no encontrada' }, HttpStatus.NOT_FOUND);
     }
 
-    // Actualizamos evento
-    // Designar un nuevo evento a la publicacion
-    if(updPubliDto.eventoNombre !== undefined){
-      const existeEvento = await this.eventoService.findOneByName(updPubliDto.eventoNombre);
-      console.log('existeEvento', existeEvento);
-      if (existeEvento) {
-        publicacion.eventoId = existeEvento._id;
-      }else{
-        throw new HttpException({ message: 'Evento no encontrado' }, HttpStatus.NOT_FOUND);
+    if(updPubliDto.eventoId !== undefined){
+      const existeEvento = await this.eventoRepository.findOne({
+        where: { id: updPubliDto.eventoId }
+      })
+      if(!existeEvento){
+        throw new HttpException({ message: `Evento con id: ${updPubliDto.eventoId}, no encontrado` }, HttpStatus.NOT_FOUND);
       }
     }
 
-    if(updPubliDto.tituloPublicacion !== undefined){
-      publicacion.tituloPublicacion = updPubliDto.tituloPublicacion;
-    }
-
-    if(updPubliDto.lugarPublicacion !== undefined){
-      publicacion.lugarPublicacion = updPubliDto.lugarPublicacion;
-    }
-
-    if(updPubliDto.fechaPublicacion !== undefined){
-      publicacion.fechaPublicacion = updPubliDto.fechaPublicacion;
-    }
-
-    return await this.publicacionRepository.save(publicacion);
+    const publicacionUpdate = Object.assign(publicacion, updPubliDto);
+    return await this.publicacionRepository.save(publicacionUpdate);
 
   }
 
-  async remove(titulo: string): Promise<PublicacionEntity> {
-    const existePublicacion = await this.publicacionRepository.findOne({ where: { tituloPublicacion: titulo } });
-    if (!existePublicacion) {
+  async updateById(id: string, updPubliDto: UpdatePublicacionDto) {
+    const publicacion = await this.publicacionRepository.findOne({ where: { id : id } });
+
+    if (!publicacion) {
       throw new HttpException({ message: 'Publicacion no encontrada' }, HttpStatus.NOT_FOUND);
+    }
+
+    if(updPubliDto.eventoId !== undefined){
+      const existeEvento = await this.eventoRepository.findOne({
+        where: { id: updPubliDto.eventoId }
+      })
+      if(!existeEvento){
+        throw new HttpException({ message: `Evento con id: ${updPubliDto.eventoId}, no encontrado` }, HttpStatus.NOT_FOUND);
+      }
+    }
+
+    const publicacionUpdate = Object.assign(publicacion, updPubliDto);
+    return await this.publicacionRepository.save(publicacionUpdate);
+
+  }
+
+  async removeById(id: string): Promise<PublicacionEntity> {
+    if(!validateUUID(id)) {
+      throw new HttpException({ message: 'Id no válido' }, HttpStatus.BAD_REQUEST);
+    }  
+    const existePublicacion = await this.publicacionRepository.findOne({ where: { id: id } });
+    if (!existePublicacion) {
+      throw new HttpException({ message: `Publicacion con id: ${id}, no encontrada` }, HttpStatus.NOT_FOUND);
     }
 
     return await this.publicacionRepository.remove(existePublicacion);
   }
 
   private async obtenerPublicacionesByIdEvento(publicacion: PublicacionEntity): Promise<any> {
-
-    const evento = await this.eventoService.findOneById(publicacion.eventoId.toString());
-
-    const categoria = await this.categoriaService.findById(evento.categoriaId.toString());
+    // console.log('publicacion', publicacion);
+    const evento = await this.eventoService.findOneById(publicacion.eventoId);
+    const categoria = await this.categoriaService.findById(evento.categoriaId);
   
 
     const eventoData = {
-      _id: evento._id,
+      id: evento.id,
       nombreEvento: evento.nombreEvento,
       fecha: evento.fecha,
       lugar: evento.lugar,
       categoria: {
-        _id: categoria._id,
+        id: categoria.id,
         nombreCategoria: categoria.nombreCategoria,
       },
     }
 
     return {
-      _id: publicacion._id,
+      id: publicacion.id,
       tituloPublicacion: publicacion.tituloPublicacion,
       lugarPublicacion: publicacion.lugarPublicacion,
       fechaPublicacion: publicacion.fechaPublicacion,
